Updated Unifact()
Rewrote Unifact() to add some concurrency. It no longer uses slices as
the primary data structure. Instead the array indices are randomly
permuted and used to populate a queue, which is popped off in pairs to
go Proc(). It appears to produce the same outputs as the original code,
but performs very slightly worse.  I suspect this is a product of
populating the queue initially and of using repeatedly using type
assertions. The type assertions can likely be optimized by using
(creating, if necessary) an IntQueue in the future.
The non-trivial overhead with launching goroutines is also at work here.
It may be that Proc() is such a basic operation that it's not worth
doing concurrently. I need to test the performance for larger Population
sizes to understand this.